+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     isFailover, getType, getTypeName, getSubtypeName, getExtraInfo, getNetworkInfo, getConnectivity
 * :sparkles: Fragmentx add findUserVisibleChildFragment, findFragmentByViewPagerCurrentItem method
 * :sparkles: Bitmapx add circularTo, circular, centerCropTo method and add related overloaded methods
+* :sparkles: Add Imagex
 
 Upgrade
 * :arrow_up: Upgrade Javax 0.6.7